backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
cryptography==43.0.1
pydantic-settings==2.5.2
python-dotenv==1.0.1
